Customize config ⚡🚀
